+++ B0.ml
@@ -33,11 +33,11 @@ let default =
     B0_meta.empty
     |> ~~ B0_meta.authors ["The webbrowser programmers"]
     |> ~~ B0_meta.maintainers ["Daniel Bünzli <daniel.buenzl i@erratique.ch>"]
-    |> ~~ B0_meta.homepage "https://erratique.ch/software/webbbrowser"
-    |> ~~ B0_meta.online_doc "https://erratique.ch/software/webbbrowser/doc/"
+    |> ~~ B0_meta.homepage "https://erratique.ch/software/webbrowser"
+    |> ~~ B0_meta.online_doc "https://erratique.ch/software/webbrowser/doc/"
     |> ~~ B0_meta.licenses ["ISC"]
-    |> ~~ B0_meta.repo "git+https://erratique.ch/repos/webbbrowser.git"
-    |> ~~ B0_meta.issues "https://github.com/dbuenzli/webbbrowser/issues"
+    |> ~~ B0_meta.repo "git+https://erratique.ch/repos/webbrowser.git"
+    |> ~~ B0_meta.issues "https://github.com/dbuenzli/webbrowser/issues"
     |> ~~ B0_meta.description_tags
       ["web"; "http"; "url"; "browser"; "cli"; "org:erratique"]
     |> B0_meta.tag B0_opam.tag
@@ -53,5 +53,5 @@ let default =
     |> ~~ B0_opam.build
       {|[["ocaml" "pkg/pkg.ml" "build" "--dev-pkg" "%{dev}%"]]|}
   in
-  B0_pack.make "default" ~doc:"webbbrowser package" ~meta ~locked:true @@
+  B0_pack.make "default" ~doc:"webbrowser package" ~meta ~locked:true @@
   B0_unit.list ()